+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2019-06-19
+
+### Added
+- Support for "&" placeholders in class names, referencing current template class.
+- Constructor method for MarkupMenuData for easily instantiating objects with data.
+
+### Changed
+- Changes to default classes and some code reordering.
+
+### Fixed
+- An issue with menu item class name output, where the result was "Array".
+
 ## [0.3.0] - 2019-06-18
 
 ### Changed